Refactor: abstract dataset and move response generation from common to dataset
